creeacion context

+++ Proyecto BASES-POO 2021/Programacion orientada a objetos/Project-POO/Project-POO/Model/Cabin.cs
@@ -16,7 +16,7 @@ namespace Project_POO
 
         public int Id { get; set; }
         public string Email { get; set; }
-        public short Phone { get; set; }
+        public int Phone { get; set; }
         public string Adress { get; set; }
 
         public virtual ICollection<Consult> Consults { get; set; }
